Move PaginationHandler to its own file
